+++ sdk-extension/opentelemetry-sdk-extension-aws/src/opentelemetry/sdk/extension/aws/trace/propagation/aws_xray_format.py
@@ -274,3 +274,12 @@ class AwsXRayFormat(TextMapPropagator):
         set_in_carrier(
             carrier, TRACE_HEADER_KEY, trace_header,
         )
+
+    def fields(self):
+        """Returns a set with the fields set in `inject`.
+
+        See
+        `opentelemetry.trace.propagation.textmap.TextMapPropagator.fields`
+        """
+
+        return {TRACE_HEADER_KEY}