+++ sdk-extension/opentelemetry-sdk-extension-aws/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add method to return fields injected by propagator
+  ([#226](https://github.com/open-telemetry/opentelemetry-python/pull/226))
+
 ## Version 0.16b0
 
 Released 2020-11-25